src/jpl.edrn.biokey.policy/pyproject.toml
[project]
name = 'jpl.edrn.biokey.policy'
dynamic = ['version']
requires-python = '>=3.11'
dependencies = [
    'jpl.wagtail.bootstrap',
    'jpl.edrn.biokey.content',
    'celery[redis]         ~= 5.2.7',
    'dj-database-url       ~= 2.1.0',
    'django                ~= 4.2.9',
    'django-auth-ldap      ~= 4.1.0',
    'django-celery-results ~= 2.5.1',
    'django-redis          ~= 5.4.0',
    'pillow                == 9.5.0',         # Must match py3-pillow package installed in Dockerfile
    'psycopg2              == 2.9.6',         # Must match py3-pyscopg2 package installed in Dockerfile
    'pymemcache            ~= 3.5.2',
    'wagtail               == 5.2.3',
    'wagtail-favicon       ~= 0.3.0',
    'wagtail-metadata      >= 4.0.1, <4.1',
    'tzdata                == 2022.1',        # Must match py3-tz-2022.1 package installed in Dockerfile
    'wagtailmenus          == 3.1.9',
    'wagtail-robots        @ git+https://github.com/unexceptable/wagtail-robots.git@9a47b6c09bec120bd5a85f4c7c1856cdc7f71f51'

]
authors = [
    {name = 'Sean Kelly', email='kelly@seankelly.biz'}
]
description = 'Site policy for BioKey, user profile management for LabCAS and the Early Detection Research Network'
readme = 'README.md'
keywords = ['user', 'profile', 'data', 'catalog', 'archive', 'edrn', 'labcas', 'cancer', 'nih', 'nci', 'jpl', 'nasa']
classifiers = [
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Healthcare Industry',
    'Intended Audience :: Science/Research',
    'License :: OSI Approved :: Apache Software License',
    'Natural Language :: English',
    'Operating System :: OS Independent',
    'Programming Language :: Python :: 3',
]
license = {file = 'LICENSE.md'}


[project.optional-dependencies]
dev = [
    'django-debug-toolbar == 3.2.2',
    'django-extensions    == 3.1.5'
]


[project.urls]
Homepage = 'https://github.com/EDRN/biokey'
Issues = 'https://github.com/EDRN/biokey/issues'


[project.scripts]


[tool.hatch.version]
path = 'src/jpl/edrn/biokey/policy/VERSION.txt'
pattern = '(?P<version>.+)'


[tool.hatch.build.targets.wheel]
packages = ['src/jpl']


[tool.hatch.metadata]
allow-direct-references = true


[build-system]
requires = ['hatchling']
build-backend = 'hatchling.build'
